experiments: Retire old atmosphere and core modules
