+++ src/eradiate/experiments/__init__.pyi
@@ -1,5 +1,5 @@
-from ._atmosphere_new import AtmosphereExperiment as AtmosphereExperiment
-from ._core_new import EarthObservationExperiment as EarthObservationExperiment
-from ._core_new import Experiment as Experiment
-from ._core_new import mi_render as mi_render
-from ._core_new import run as run
+from ._atmosphere import AtmosphereExperiment as AtmosphereExperiment
+from ._core import EarthObservationExperiment as EarthObservationExperiment
+from ._core import Experiment as Experiment
+from ._core import mi_render as mi_render
+from ._core import run as run
