+++ tests/02_eradiate/01_unit/experiments/test_core.py
